move more classes out of ExpDict.*
